code/keyword_extraction/UIE/requirements.txt
-i https://opentuna.cn/pypi/web/simple/
paddlepaddle-gpu
paddlenlp
